Add air_chiller, refrigeration, and fan_on_off tests
This adds tests for air_chillers, fan_on_off, and refrigeration_system.

The air_chillers and refrigeration_system tests are not best practice
systems but rather setup to test all or most possible combinations.

This also includes a change to allow the use of 10.9 sdk and target as
well as Ruby 2.0