Run with 2.9.1 official, and add ShadowCalculation OSM test for it. I decided 2.9.1 made sense, read below
`ShadowCalculation` was added a long time ago, but the 'Calculation Method' (now called 'Shading Calculation Update Frequency Method' was added in v2.9.0, and since it's closed to 2.9.0, might as well just do 2.9.1.
I didn't think it was worth it to really go back to version 2.0.5 for that, this'll do just fine.